chore: Update routes in CI/CD workflow
This is done because the let's encrypt cert authority keeps rate limiting the deployments.
It seems that the TLS certs are regenerated with each deployment.
Instad of oscar.stg.cloudflavor.io, it will be oscar-stg-35bae4e1.cloudflavor.workers.dev
for example. These routes are handled by cloudflare and still have TLS.

Signed-off-by: Victor Palade <victor@cloudflavor.io>