commands: introduce new commands
These changes introduce restart-workflow, stop-workflow, cancel-workflow
close, reopen, merge, rename.
It also adds a new event for workflow_run failed. If a failed job is detected for the main event, then a new issue is created by the bot that references it.
This new issue will be labeled with failed-ci for now.

Signed-off-by: Victor Palade <victor@cloudflavor.io>