config: add permissions config file for the app
The env variable OSCAR_ACCESS_CONFIG_URI is used to specify the URI of the
permissions configuration file. The file is in TOML format and for the
time being contains a single entry that specify PI-Victor is the admin.
No other users will be able to execute commands for the bot now.
Closes: https://github.com/cloudflavor/oscar/issues/15

Signed-off-by: Victor Palade <victor@cloudflavor.io>