feat:  Merge on successful jobs
Before merging the bot will check if the jobs are successful and in completed status.
Otherwise it will refuse to merge them.

This commit also adds a new command, '/merge force', to the handlePrMergeCommand function. When this command is used, the pull request will be force merged, bypassing checks and conflicts. A comment will be added to indicate the force merge.

Signed-off-by: Victor Palade <victor@cloudflavor.io>