feat: Add more commands.
lock/unlock - lock, unlock an issue.
milestone - add a milestone. mileston clear will remove it.
pin/unpin - pin/unpin issue. had to use graphql since rest does not seem to provide a way to update the issue for pinning.

draft - move a PR to draft. this only works on payed plans.
approve/unapprove - does what it says on the thin. It also adds more clauses to merge command, like not merging if it's already merged, if it's a draft, if it has conflicts, etc.

Signed-off-by: Victor Palade <victor@cloudflavor.io>